+++ csatar-plugins/csatar/lang/en/lang.php
@@ -423,7 +423,11 @@
             ],
             'gallery' => [
                 'gallery' => 'Gallery'
-            ]
+            ],
+            'permissions' => [
+                'permissions' => 'Permissions',
+                'allPermissionsForScout' => 'All permissions for every model and field for Scout mandate.'
+            ],
         ],
         'component' => [
             'general' => [