+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- CS-268 - Added new pictograms, and changed definition of existing ones
 - CS-229 - Changed grid solution using cards
 - CS-229 - Deleted earlier grid solution from design system
 - CS-229 - Fixed cards bottom rounded corner