+++ Changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- CS-47 Created basics of the rights system
 - CS-268 - Added new pictograms, and changed definition of existing ones
 - CS-199 - Extended social login plugin to handle Scout and User model connections
 - CS-229 - Changed grid solution using cards